serial: sandbox: use ssize_t to match os_read
Acked-by: Simon Glass <sjg@chromium.org>

+++ drivers/serial/sandbox.c
@@ -50,7 +50,7 @@ void serial_puts(const char *str)
 int serial_getc(void)
 {
 	char buf;
-	int count;
+	ssize_t count;
 
 	count = os_read(0, &buf, 1);
 	return count == 1 ? buf : 0;
